BUGFIX: call clearscreen() during initial game sync up
Initialize screen related variables on initial startup (which is normally
done after receiving CL_TOK in normal play so should be done when we
manually wait for initial CL_TOK). There is probably a variable that
should be initialized on startup that isn't. But this ensures
everything is set correctly. Failure to do this causes illchar
response message to appear in wrong part of screen which eventually corrupts
game play.